separate lightning model from train file
