+++ model.py
@@ -1,6 +1,5 @@
 import lightning as L
 import torch
-import torch.nn.functional as F
 import torch.nn as nn
